Reorder ShopperInteraction Enum and correct default
The default value for the ShopperInteraction is Ecommerce according
to the docs https://docs.adyen.com/api-explorer/#/Payment/v40/authorise.
However the default value for an Enum is 0 which, if no
values are specified, corresponds to the first enum value
https://stackoverflow.com/questions/529929/choosing-the-default-value-of-an-enum-type-without-having-to-change-values/529937.
This along with the EmitDefaultValue = false means that if ShopperInteraction.Moto
was passed in the PaymentRequest object it would be seen as default value
and left out of the serialization. When the API call is made the lack of
a ShopperInteraction means the actual default value is used which is Ecommerce.
Thus all Moto payments actually get processed as Ecommerce.

Reordered ShopperInteraction so Ecommerce is the default.
Added tests to confirm that reordering enum allows ShopperInteraction.Moto
to be correctly serialized.