Update HMACValidator.cs
notificationRequestItem Success is a boolean and will become True or False starting with uppercase.
While in the json is  success:true of success:false with lowercase.
this results in a different hash. and an false IsValidHmac
this is something I ran in to while parsing the jsonrequest in the NotificationHandler.HandleNotificationRequest(jsonRequest).
maybe this is an error on my part, but I thought I should let you know and don't know if the proposed ToLower is the best solution?