Update HMACValidator.cs
In the example on the website the boolean is shown with a lowercase t. 
I presume (maybe mistakenly) that the encyption done on the adyenside is done with a lower case t(rue) or f(alse)
doing this in dotnet will result in a string with an uppercase T or F.  (that why is propose a ToLower()
rendering a different hash and will give a false negative.