a proposed change to GetDataToSign function in the HMACvalidator (#150)
* Update UtilTest.cs

Shouldn't this be the same as in the java test?      
because your are using the same data and same key? 
in the java test: 
assertEquals("pspReference:originalReference:merchantAccount:reference:1000:EUR:EVENT:true", data);

* Update HMACValidator.cs

In the example on the website the boolean is shown with a lowercase t. 
I presume (maybe mistakenly) that the encyption done on the adyenside is done with a lower case t(rue) or f(alse)
doing this in dotnet will result in a string with an uppercase T or F.  (that why is propose a ToLower()
rendering a different hash and will give a false negative.

* Update HMACValidator.cs